% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dove3.R
\name{dove3}
\alias{dove3}
\title{Effectiveness of Vaccination and Prior Infection}
\usage{
dove3(
  formula,
  data,
  vaccine_infection_interaction = FALSE,
  vaccine_knots = NULL,
  vaccine_uninfected_knots = NULL,
  vaccine_infected_knots = NULL,
  prior_infection_knots = NULL,
  related_vaccine_types = NULL,
  last_piece_constant = FALSE,
  reinfection_cutoff = 14,
  plots = TRUE
)
}
\arguments{
\item{formula}{A formula object, with the response on the left-hand side of a
'~' operator, and the covariates and exposure() function on the right-hand side.
The response contains the outcome information and
must be specified through function 'outcome()'.
The exposure() function must be used to specify the information of each exposure.
See ?exposure, ?outcome, and Details for further information.}

\item{data}{A data.frame object. The data.frame in which to interpret the
variable names in formula. Must contain the subject ID,
the entry and censoring times, the event time,
the vaccination time and vaccine type, the infection time and
infection type (determined by the dominant variant),
and the covariates. See Details.}

\item{vaccine_infection_interaction}{A logical object. If TRUE, interaction term
between vaccination and prior infection is included in the model. That is,
vaccine effects are allowed to be different between previously uninfected subjects
and previously infected subjects. If FALSE (default), there is no interaction
term in the model, i.e., average vaccine effects among all subjects are estimated,
regardless of the prior infection status.}

\item{vaccine_knots}{A list object or NULL. The \eqn{k}th element specifies
the knots of the piecewise linear function for the log rate/hazard ratio
of the \eqn{k}th vaccine type. If NULL, knots will be placed
at every month by default.
This input is ignored when vaccine_infection_interaction = TRUE.}

\item{vaccine_uninfected_knots}{A list object or NULL.
The \eqn{k}th element specifies the knots of the piecewise linear function
for the log rate/hazard ratio of the \eqn{k}th vaccine type
given no prior infection before vaccination.
If NULL, knots will be placed at every month by default.
This input is ignored when vaccine_infection_interaction = FALSE.}

\item{vaccine_infected_knots}{A list object or NULL.
The \eqn{k}th element specifies the knots of the piecewise linear function
for the log rate/hazard ratio of the \eqn{k}th vaccine type
given at least one prior infection before vaccination.
If NULL, knots will be placed at every other month by default.
This input is ignored when vaccine_infection_interaction = FALSE.}

\item{prior_infection_knots}{A list object or NULL.
The \eqn{k}th element specifies the knots of the piecewise linear function
for the log rate/hazard ratio of the \eqn{k}th infection type.
The first knot should be placed at reinfection_cutoff.
If NULL, a default set of knots will be used, with the first knot placed
at reinfection_cutoff and the subsequent knots placed at every three months.}

\item{related_vaccine_types}{A list object or NULL.
Each element of the list takes the form c(i, j) and imposes a constraint that the
slope of the first piece of the piecewise linear function is the same between
the ith and jth vaccine types.
This input is intended to account for the fact that the 1-dose and 2-dose
regimens of an mRNA vaccine should have the same effectiveness from
the receipt of the first dose until the receipt of the second dose.
If NULL (default), there is no constraint on the piecewise linear functions
for different vaccine types during the estimation.}

\item{last_piece_constant}{A logical object. If FALSE (default), effectiveness
is allowed to vary after the last knot. If TRUE, VE is assumed to be
constant after the last knot.}

\item{reinfection_cutoff}{A positive scalar object. Positive results at two time
points separated by a gap larger than this number (in days) are considered as
two different infections.}

\item{plots}{A logical object. If TRUE (default), plots of the estimated
effectiveness of each vaccine type and prior infection type against the
clinical outcome of interest and their 95\% confidence intervals will be
automatically generated. If FALSE, plots will not be generated.}
}
\value{
An list object with elements

  \item{covariates}{A matrix containing the estimated (log) hazard ratio of
    each covariate, together with the estimated standard error, the 95\%
    confidence interval, and the two-sided p-value for testing no covariate
    effect.}

  \item{effectiveness}{A list of matrices, one for each type of exposure.
    Each matrix contains the daily effectiveness estimates in reducing the rate
    or hazard of the clinical outcome of interest, together with the
    standard errors and the 95\% confidence intervals.}

  \item{plots}{A list of plot objects returned by ggplot().}
}
\description{
Estimates the potentially waning long-term effectiveness of vaccination and
 prior infection against COVID-19 outcomes in observational studies.
 Effects of all exposures (i.e., vaccination and prior infection) are estimated
 simutaneously under a single Cox regression model, allowing the outcome
 of interest to be a recurrent event.
}
\details{
The log intensity or hazard ratio for an exposure is a piecewise
 linear function of time since the exposure.
 Specific constraint on the first and last piece is allowed.
 In addition, vaccine effects can be evaluated among previously uninfected
 and infected subjects separately.

The information required for an analysis is
  \describe{
    \item{Subject ID:}{Number that identifies subjects.}
    \item{Entry Time:}{Calendar time when the subject enters the risk set.}
    \item{Event Time:}{Calendar time when the subject experiences
      the clinical outcome of interest (e.g., SARS-CoV-2 Infection,
      hospitalization, or death caused by infection), with NA, Inf, or an
      arbitrary value greater than the censoring time if the subject does not
      experience an event.}
    \item{Censoring Time:}{Calendar time when the subject moves out of the risk set.}
    \item{Vaccination Time:}{Calendar time when vaccination takes place,
      with NA, Inf, or an arbitrary value greater than the study end time
      if the subject is not vaccinated during the study period.}
    \item{Vaccine Type:}{Categorical variable indicating which vaccine type
      the subject receives. Must be an integer between 1 and the total number
      of vaccine types, with NA or an arbitrary value within this range if the
      subject is not vaccinated during the study period.}
    \item{Infection Time:}{Calendar time when the subject is infected,
      with NA, Inf, or an arbitrary value that is larger than the study end
      time if the subject is not infected during the study period.}
   \item{Infection Type:}{Categorical variable indicating which dominant variant
      the infection is associated with. Must be an integer between 1 and the total
      number of variants under investigation, with NA or an arbitrary value within
      this range if the subject is not infected during the study period.}
    \item{Covariates:}{Baseline covariates (e.g., age, gender, ethnicity).}
   }

Note that a subject can have multiple rows and each row
corresponds to an infection.
The event_time is the time of the clinical outcome of interest caused by the
infection in the same row, with NA, Inf, or an
arbitrary value greater than the censor_time if that infection does not cause
the outcome.
Rows from the same subject are sorted by the infection_time.
If a subject never experiences the clinical outcome of interest during the
entire study period, this subject will have only one record, with the event_time
being NA, Inf, or an arbitrary value greater than the censor_time, and with the
infection_time being NA, Inf, or any arbitrary value.

All the time variables are measured from the same time origin
and are specified in units of days.
For each individual, the entry_time and censor_time must satisfy
entry_time \eqn{\le} censor_time.
If entry_time > censor_time, the case will be
removed from the analysis and a message will be generated.
The definitions of entry_time and censor_time depend on specific analyses.
For example, entry_time is set to be 0 for all subjects if effectiveness of primary
vaccine series is of interest, and is set to be the completion time of the
primary vaccine series if effectiveness of boosters is of interest.
Thus, the entry_time and censor_time are not necessarily the beginning and
end of the follow-up period for a subject, and it is possible that
event_time is smaller than entry_time or larger than censor_time.

All the categorical variables (i.e., subject_id, vaccine_type, infection_type)
must be labelled as 1, 2, \eqn{\dots}, \eqn{K}, where \eqn{K} is the total
number of categories.


The general structure of the formula input is
  \preformatted{
  outcome(subject_id, entry_time, event_time, censor_time) ~ covariates +
    exposure(vaccine_time, vaccine_type, infection_time, infection_type)
  }

The response variable contains the information of event time and event status
for each infection and must be specified through function 'outcome()'.
All the variables related to vaccination and infection are specified
through function 'exposure()'.
The covariates can be either numerical or categorical.
If categorical covariates are provided, all other categories are compared to
the first category.
}
\examples{
data(exampleData)

# specify the knots for each exposure
vaccine.knots = list("vac.type1" = c(30),
                     "vac.type2" = c(30,60),
                     "vac.type3" = c(30,60))
prior.infection.knots = list("inf.type1" = c(14, 120),
                             "inf.type2" = c(14,120))

# Fit the simple model without interaction or related vaccine types
fit1 <- dove3(formula = outcome(subject.id, entry.time, event.time, censor.time) ~
                age + gender + priority +
                exposure(Vtime, Vtype, infection.time, infection.type),
      data = exampleData,
      vaccine_knots = vaccine.knots,
      prior_infection_knots = prior.infection.knots)

# Specify the knots for vaccination without and with prior infection
vaccine.uninfected.knots = list("vac.noinf.type1" = c(30),
                                "vac.noinf.type2" = c(30,60),
                                "vac.noinf.type3" = c(30,60))

vaccine.infected.knots = list("vac.noinf.type1" = c(30),
                                "vac.noinf.type2" = c(60),
                                "vac.noinf.type3" = c(60))

# Fit the model with interaction between vaccination and prior infection status,
# and impose a constraint on the first pieces of the first two vaccine types.
fit2 <- dove3(formula = outcome(subject.id, entry.time, event.time, censor.time) ~
                age + gender + priority +
                exposure(Vtime, Vtype, infection.time, infection.type),
      data = exampleData,
      vaccine_infection_interaction = TRUE,
      vaccine_uninfected_knots = vaccine.uninfected.knots,
      vaccine_infected_knots = vaccine.infected.knots,
      prior_infection_knots = prior.infection.knots,
      related_vaccine_types = list(c(1,2)))

}
\references{
Lin D, Gu Y, Xu Y, et al. Association of Primary and Booster
  Vaccination and Prior Infection With SARS-CoV-2 Infection and Severe
  COVID-19 Outcomes. JAMA. Published online September 26, 2022.
  doi:10.1001/jama.2022.17876
}
