% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{outcome}
\alias{outcome}
\title{Specify Outcome Variables}
\usage{
outcome(subject_id, entry_time, event_time, censor_time)
}
\arguments{
\item{subject_id}{The variable for the subject ID that is used to identify
records from the same subject, must be an integer between 1 and the total
number of subjects.}

\item{entry_time}{The variable for the time when the subject enters the risk set.
For example, entry_time is set to be 0 for all subjects if effectiveness of
primary vaccine series is of interest, and is set to be the time of
primary vaccine series if effectiveness of boosters is of interest.}

\item{event_time}{The variable for the time when the subject has an event.}

\item{censor_time}{The variable for the time when the subject is censored.
For example, subjects are censored when they receive boosters if
effectiveness of primary vaccine series is of interest.
Likewise, subjects are censored when they receive second boosters if
effectiveness of first boosters is of interest.}
}
\value{
This function is intended to be used only in the model statement
 of dove3(). The result, a matrix, is used internally.
}
\description{
This function is used in the model statement of dove3() to specify
 the subject id, entry time, event time, and censoring time.
}
